Refactor session migration logic into separate utility function
